feat: Implement Fees module with fee preview functionality and integrate into AppModule
