+++ src/modules/fees/routes/entities/fee-preview-transaction.dto.entity.ts
@@ -2,7 +2,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import type { Address, Hex } from 'viem';
 import { Operation } from '@/modules/safe/domain/entities/operation.entity';
-import { FeePreviewTransactionDtoSchema } from '@/modules/transactions/routes/entities/schemas/fee-preview-transaction.dto.schema';
+import { FeePreviewTransactionDtoSchema } from '@/modules/fees/routes/entities/schemas/fee-preview-transaction.dto.schema';
 import { z } from 'zod';
 
 export class FeePreviewTransactionDto implements z.infer<
