+++ src/modules/transactions/transactions.module.ts
@@ -61,7 +61,6 @@ import { BridgeTransactionMapper } from '@/modules/transactions/routes/mappers/c
 import { LiFiDecoderModule } from '@/modules/bridge/domain/contracts/decoders/lifi-decoder.helper';
 import { LiFiHelperModule } from '@/modules/transactions/routes/helpers/lifi-helper';
 import { BridgeModule } from '@/modules/bridge/bridge.module';
-import { FeeServiceApiModule } from '@/datasources/fee-service-api/fee-service-api.module';
 
 @Module({
   controllers: [TransactionsController],
@@ -73,7 +72,6 @@ import { FeeServiceApiModule } from '@/datasources/fee-service-api/fee-service-a
     DataDecoderModule,
     DelegatesV2RepositoryModule,
     EarnModule,
-    FeeServiceApiModule,
     GPv2DecoderModule,
     HumanDescriptionModule,
     KilnNativeStakingHelperModule,
